[Modify] 수상 요소 수정 및 삭제 시 본인확인

+++ back/src/routers/awardRouter.js
@@ -66,6 +66,7 @@ awardRouter.put(
         try {
             // URI로부터 수상 요소 id를 추출함.
             const award_id = req.params.id;
+            const user_id = req.currentUserId;
             const toUpdate = req.toUpdate;
             console.log('toUpdate : ', toUpdate);
             if (Object.keys(toUpdate).length === 0) {
@@ -75,6 +76,7 @@ awardRouter.put(
             // 해당 수상 요소 아이디로 수상 요소 정보를 db에서 찾아 업데이트함. 업데이트 요소가 없을 시 생략함
             const updatedAward = await AwardService.setAward({
                 award_id,
+                user_id,
                 toUpdate
             });
 
@@ -92,7 +94,8 @@ awardRouter.delete(
         try {
             // URI로부터 수상 요소 id를 추출함.
             const award_id = req.params.id;
-            await AwardService.deleteAward({ award_id });
+            const user_id = req.currentUserId;
+            await AwardService.deleteAward({ award_id, user_id });
             console.log(award_id);
             res.status(200).send('삭제되었습니다.');
         } catch (error) {
